fix: goalsList storys working
